Fix formatting of IPv6 LB VIPs
I encountered the following error in an ovn-controller log:

019-11-04T14:15:02Z|00025|lflow|WARN|error parsing actions "ct_lb(2600:1f16:d2d:3003:9dbc:39cc:ef6c:6dbb:6443);": Invalid numeric constant.

The root cause is that the IPv6 address was not wrapped in square
brackets when ovn-kubernetes configured the VIPs for a load balancer.

This patch fixes this issue, along with a number of other places where
net.JoinHostPort() needs to be used instead of performing the same
operation manually in a way that works for IPv4 only.