Don't panic when no API endpoint is ready
When running in an environment where the only address for the
kubernetes.default API endpoint is IPv6, I was seeing a panic in this
function because there's a log message that assumes there is at least
one address.  I think what's happening is that in some cases, the IPv6
address isn't ready yet, so Addresses on the endpoint is empty.  In
that case, just bail early and wait for the address to become ready
and listed in Addresses.